Add black formatter for Python files & code_format_everything can format only staged changes

Closes #790

See merge request baci/baci!1146

+++ tests/post_processing_test/readutils.py
@@ -10,19 +10,19 @@ def read_string(file_handle, length):
         result_str[pos] = file_handle.read(1)
         pos += 1
 
-    return b''.join(result_str).decode("ascii")
+    return b"".join(result_str).decode("ascii")
 
 
 def read_int(file_handle):
     """Reads integer"""
-    return struct.unpack('<i', file_handle.read(4))[0]
+    return struct.unpack("<i", file_handle.read(4))[0]
 
 
 def read_ints(file_handle, length):
     """Reads integer"""
-    return np.fromfile(file_handle, dtype='<i', count=length)
+    return np.fromfile(file_handle, dtype="<i", count=length)
 
 
 def read_floats(file_handle, length):
     """reads array of length floats"""
-    return np.fromfile(file_handle, dtype='<f4', count=length)
+    return np.fromfile(file_handle, dtype="<f4", count=length)
